Revert "Multiple choice"
