+++ src/main/java/JpaApplication/Question.java
@@ -1,6 +1,4 @@
-package JpaApplication.Model;
-
-import JpaApplication.Model.Answers.Answer;
+package JpaApplication;
 
 import javax.persistence.*;
 import java.util.ArrayList;
@@ -14,7 +12,6 @@ public class Question {
     private long id;
 
     private String questions = "";
-    private String questionType = "";
 
     @OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL, mappedBy = "question")
     private List<Answer> answers;
@@ -48,6 +45,15 @@ public class Question {
         this.answers = answers;
     }
 
+
+    public void setAnswersByType(String type) {
+        if(type.equals("Open-Ended")) {
+            answers = new ArrayList<Answer>();
+        }else {
+            answers = new ArrayList<Answer>();
+        }
+    }
+
     public Answer getAnswer(int index) {
         return answers.get(index);
     }
@@ -70,14 +76,6 @@ public class Question {
 
     public void setQuestions(String question) {
         this.questions = question;
-    }
-
-    public String getQuestionType() {
-        return questionType;
-    }
-
-    public void setQuestionType(String questionType) {
-        this.questionType = questionType;
     }
 
     @Override
