+++ src/main/java/JpaApplication/QuestionController.java
@@ -1,8 +1,11 @@
-package JpaApplication.Controller;
+package JpaApplication;
 
-import JpaApplication.Model.Question;
-import JpaApplication.Repository.QuestionRepository;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
