+++ src/main/java/JpaApplication/SurveyController.java
@@ -1,12 +1,5 @@
-package JpaApplication.Controller;
+package JpaApplication;
 
-import JpaApplication.Model.Answers.Answer;
-import JpaApplication.Model.Answers.MultipleChoice;
-import JpaApplication.Model.Answers.OpenEnded;
-import JpaApplication.Model.Question;
-import JpaApplication.Model.Survey;
-import JpaApplication.Repository.QuestionRepository;
-import JpaApplication.Repository.SurveyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -33,17 +26,14 @@ public class SurveyController {
     @PostMapping("/addQuestion")
     public String addQuestionSubmit(@ModelAttribute Question question, Model model) {
         model.addAttribute("question", question);
-        MultipleChoice answer = new MultipleChoice("answer1", "answer2", "answer3", "answer4");
-        Answer answer1 = new OpenEnded("answer");
+        Answer answer = new Answer("answer");
+        question.setAnswersByType("Open-Ended");
         question.addAnswer(answer);
-        question.addAnswer(answer1);
+//
         QuestionRepo.save(question);
 
-//      Tests
         Question q1 = QuestionRepo.findByQuestions(question.getQuestions());
         System.out.println(q1);
-        System.out.println(QuestionRepo.findAll());
-        System.out.println(q1.getAnswers());
 
         return "result";
     }
