+++ src/test/java/JpaApplication/UserTest.java
@@ -1,8 +1,6 @@
 package JpaApplication;
 
 import static org.junit.Assert.*;
-
-import JpaApplication.Model.User;
 import org.junit.Test;
 
 public class UserTest {
